% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cond.R
\name{cond_imp_lm}
\alias{cond_imp_lm}
\title{Conditional Imputation for Linear Scalar-on-Funciton Regression}
\usage{
cond_imp_lm(dat, workGrid, k = 5, seed = NULL, impute_type = "Multiple",
  muy = NULL, var_y = NULL, Cxy, var_delt = NULL, mux = NULL,
  Cx = NULL, phi = NULL, lam = NULL, tol = 1e-05)
}
\arguments{
\item{dat}{An \eqn{n \times 4} data frame (where \eqn{N} is the number of subjects,
each with \eqn{m_i} observations, so that \eqn{\sum_{i=1}^N m_i = n})
expected to have variables 'X','y','subj', and 'argvals'.}

\item{workGrid}{A vector of the unique desired grid points on which to evaluate the function.
The length of this vector will be called \eqn{M}.}

\item{k}{An integer specifying the number of desired imputations, if \code{impute_type} is "Multiple".}

\item{seed}{A numeric value used to set the seed for reproducibility (useful for multiple imputation).}

\item{impute_type}{A string used to choose between mean and multiple imputation. Should be one of "Mean"
or "Multiple".}

\item{muy}{A single numeric value representing the mean of the response variable, \eqn{Y}.}

\item{var_y}{A single numeric value representing the variance of the response variable, \eqn{Y}.}

\item{Cxy}{A numeric vector of length \code{M} specifying the cross covariance between \eqn{X(t)} and \eqn{Y}.}

\item{var_delt}{A number representing \eqn{\sigma^2_\delta}, the variance of the noise.}

\item{mux}{A numeric vector of length \code{M} specifying the mean function for \eqn{X(t)}, evaluated at \code{workGrid}.}

\item{Cx}{An \eqn{M \times M} matrix for the covariance function of \eqn{X(t)}, evaluated at \code{workGrid}.}

\item{phi}{An \eqn{M \times J} matrix whose columns are the \eqn{J} eigenfunctions of \eqn{C_X}, each evaluate at \code{workGrid}.}

\item{lam}{A length-\eqn{J} numeric vector containing the eigenvalues of \eqn{C_X}.}

\item{tol}{A (small) numerical value that sets the tolerance for trimming the eigenvalues of the conditional covariance.}
}
\value{
Either a \eqn{N\times J} matrix of imputed scores if \code{impute_type} is set to "Mean", or a 3-dimensional array
of dimension \eqn{N\times J\times k} if \code{impute_type} is set to "Multiple".
}
\description{
Given the imputation parameters, this function imputes the scores using
their distribution conditional on the observed values of the curves and
the response variable. Muliple or mean imputation can be chosen.
}
\details{
The variables of \code{dat} should be specified as follows: 'X' specifies the observed values of the curves
(no missing values here); 'y' should be a vector of numeric values such that the \eqn{m_i} values of 'y'
are the same for each subject; 'subj' should contain unique numeric identifiers for each subject,
and 'argvals' should indicate the time point at which each observation was made
(note that these values should be a subset of \code{workGrid}).
}
\author{
Jusitn Petrovich, \email{jpetrovich02@gmail.com}
}
