% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misfit.R
\name{misfit}
\alias{misfit}
\title{MISFIT (Multiple Imputation for Sparsely-sampled Functions at Irregular Times)}
\usage{
misfit(dat, grid, K, J, family = "Gaussian")
}
\description{
MISFIT (Multiple Imputation for Sparsely-sampled Functions at Irregular Times)
}
