% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_est_linear.R
\name{param_est_linear}
\alias{param_est_linear}
\title{Imputation Parameter Estimation for Linear Regression}
\usage{
param_est_linear(
  dat,
  y,
  workGrid,
  M,
  cond.y = TRUE,
  use_fcr = TRUE,
  fcr.args = list(use_bam = T, niter = 1),
  k = -1,
  face.args = list(knots = 12, lower = -3, pve = 0.95),
  FPCA.args = NULL
)
}
\arguments{
\item{dat}{A data frame with \eqn{n} rows (where \eqn{N} is the number of subjects,
each with \eqn{m_i} observations, so that \eqn{\sum_{i=1}^N m_i = n})
expected to have either 3 or 4. If \code{cond.y} is TRUE, should include
4 columns, with variables 'X','y','subj', and 'argvals'. If \code{cond.y}
is FALSE, only 3 columns are needed (no 'y' variable is used).}

\item{workGrid}{A length \eqn{M} vector of the unique desired grid points on which to evaluate the function.}

\item{cond.y}{A logical indicator to determine whether imputation will be done conditional on \eqn{Y}.}

\item{fcr.args}{A list of arguments to be passed to the underlying function \code{fcr}.}

\item{k}{Dimension of the smooth terms used in \code{fcr}. Only needs to be specified if \code{cond.y} is TRUE.}

\item{face.args}{A list of arguments to be passed to the underlying function \code{face.sparse}.}

\item{nPhi}{An integer value, indicating the number of random effects to include in the model. This is
passed to \code{fcr} and is only used when \code{cond.y} is TRUE. See \code{fcr} for more details.}
}
\value{

}
\description{
Given the data, this function estimates the imputation parameters.
This function is a wrapper for \code{fcr} if \code{cond.y} is set to TRUE,
and \code{face.sparse} if \code{cond.y} is set to FALSE.
}
\references{

}
\author{

}
