% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misfit.R
\name{misfit}
\alias{misfit}
\title{MISFIT (Multiple Imputation for Sparsely-sampled Functions at Irregular Times)}
\usage{
misfit(dat, grid, K = 10, J, family = "Gaussian", seed, ret_allxi = F)
}
\description{
Performs MISFIT for either linear (\code{family="Gaussian"}) or logistic
(\code{family="Binomial"}) regression.
}
